EY-4499 Flytte over mer til brev model så brevkafka kan lese dette
