+++ libs/etterlatte-brev-model/build.gradle.kts
@@ -5,4 +5,6 @@ plugins {
 dependencies {
     implementation(project(":libs:saksbehandling-common"))
     implementation("no.nav.pensjon.brevbaker:brevbaker-api-model-common:1.4.0")
+
+    compileOnly(libs.logging.slf4japi)
 }
